+++ .monorepolint.config.mjs
@@ -17,6 +17,7 @@ const nonStandardPackages = [
   "mytsup",
   "tsconfig",
   "@osdk/examples.todoapp",
+  "@osdk/tests.*",
 ];
 
 const cache = new Map();
@@ -132,7 +133,19 @@ function standardPackageRules(shared) {
           publishConfig: {
             "access": "public",
           },
-          files: ["build/types", "build/js", "CHANGELOG.md", "package.json"],
+          files: [
+            "build/types",
+            "build/js",
+            "CHANGELOG.md",
+            "package.json",
+
+            // fallback entries for "submodule imports" in legacy projects
+            "*.d.ts",
+          ],
+
+          main: "./build/js/index.js",
+          module: "./build/js/index.mjs",
+          types: "./build/types/index.d.ts",
         },
       },
     }),
@@ -248,6 +261,7 @@ export default {
         entries: {
           license: "Apache-2.0",
         },
+        entriesExist: ["version"],
       },
     }),
 
@@ -257,7 +271,7 @@ export default {
           private: true,
         },
       },
-      includePackages: ["@osdk/example.*"],
+      includePackages: ["@osdk/example.*", "@osdk/tests.*"],
     }),
 
     alphabeticalDependencies({ includeWorkspaceRoot: true }),
@@ -281,6 +295,12 @@ export default {
           "publishConfig",
           "imports",
           "keywords",
+          "bin",
+          "files",
+          // since these are just for fallback support we can drop to bottom
+          "main",
+          "module",
+          "types",
         ],
       },
     }),