+++ packages/generator/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @osdk/generator
 
+## 0.0.16
+
+### Patch Changes
+
+- ed78694: Fixes imports from older module resolution
+- Updated dependencies [ed78694]
+  - @osdk/gateway@0.0.10
+  - @osdk/api@0.0.17
+
+## 0.0.15
+
+### Patch Changes
+
+- b59cb10: Adding backcompatibility for internal types
+
 ## 0.0.14
 
 ### Patch Changes